v3

+++ 2. Ejercicio 2.py
@@ -1,5 +1,5 @@
-base = float(input("Ingrese la base: "))
-altura = float(input("Ingrese la altura: "))
+base = int(input("Ingrese la base: "))
+altura = int(input("Ingrese la altura: "))
 area = base * altura
 
 print("El area del rectangulo es:", area)
